Extract d3 code into a hook
